Move BreakDownClassifier into ProgressReporterUtils (former ProgressReporter$Utils)
