+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/ProgressReporter.java
@@ -32,7 +32,6 @@ import java.lang.management.GarbageCollectorMXBean;
 import java.lang.management.ManagementFactory;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.security.CodeSource;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Comparator;
@@ -54,8 +53,6 @@ import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import com.oracle.svm.core.configure.ConditionalRuntimeValue;
-import com.oracle.svm.core.jdk.resources.ResourceStorageEntryBase;
 import org.graalvm.nativeimage.ImageSingletons;
 import org.graalvm.nativeimage.hosted.Feature;
 import org.graalvm.nativeimage.impl.ImageSingletonsSupport;
@@ -75,9 +72,11 @@ import com.oracle.svm.core.SubstrateGCOptions;
 import com.oracle.svm.core.SubstrateOptions;
 import com.oracle.svm.core.SubstrateUtil;
 import com.oracle.svm.core.VM;
+import com.oracle.svm.core.configure.ConditionalRuntimeValue;
 import com.oracle.svm.core.heap.Heap;
 import com.oracle.svm.core.hub.ClassForNameSupport;
 import com.oracle.svm.core.jdk.Resources;
+import com.oracle.svm.core.jdk.resources.ResourceStorageEntryBase;
 import com.oracle.svm.core.option.AccumulatingLocatableMultiOptionValue;
 import com.oracle.svm.core.option.HostedOptionKey;
 import com.oracle.svm.core.option.HostedOptionValues;
@@ -116,7 +115,7 @@ import jdk.graal.compiler.util.json.JsonWriter;
 
 @SingletonTraits(access = BuildtimeAccessOnly.class, layeredCallbacks = NoLayeredCallbacks.class, layeredInstallationKind = Independent.class)
 public class ProgressReporter {
-    private static final int CHARACTERS_PER_LINE;
+    static final int CHARACTERS_PER_LINE;
     private static final String HEADLINE_SEPARATOR;
     private static final String LINE_SEPARATOR;
     private static final int MAX_NUM_BREAKDOWN = 10;
@@ -178,8 +177,8 @@ public class ProgressReporter {
 
     static {
         CHARACTERS_PER_LINE = SubstrateUtil.isNonInteractiveTerminal() ? ProgressReporterCHelper.MAX_CHARACTERS_PER_LINE : ProgressReporterCHelper.getTerminalWindowColumnsClamped();
-        HEADLINE_SEPARATOR = Utils.stringFilledWith(CHARACTERS_PER_LINE, "=");
-        LINE_SEPARATOR = Utils.stringFilledWith(CHARACTERS_PER_LINE, "-");
+        HEADLINE_SEPARATOR = ProgressReporterUtils.stringFilledWith(CHARACTERS_PER_LINE, "=");
+        LINE_SEPARATOR = ProgressReporterUtils.stringFilledWith(CHARACTERS_PER_LINE, "-");
     }
 
     public static ProgressReporter singleton() {
@@ -444,9 +443,9 @@ public class ProgressReporter {
 
         l().printLineSeparator();
         l().yellowBold().doclink("Build resources", "#glossary-build-resources").a(":").reset().println();
-        l().a(" - %s of memory (%.1f%% of system memory, %s)", ByteFormattingUtil.bytesToHuman(maxMemory), Utils.toPercentage(maxMemory, totalMemorySize), memoryUsageReason).println();
+        l().a(" - %s of memory (%.1f%% of system memory, %s)", ByteFormattingUtil.bytesToHuman(maxMemory), ProgressReporterUtils.toPercentage(maxMemory, totalMemorySize), memoryUsageReason).println();
         l().a(" - %s thread(s) (%.1f%% of %s available processor(s), %s)",
-                        maxNumberOfThreads, Utils.toPercentage(maxNumberOfThreads, availableProcessors), availableProcessors, maxNumberOfThreadsSuffix).println();
+                        maxNumberOfThreads, ProgressReporterUtils.toPercentage(maxNumberOfThreads, availableProcessors), availableProcessors, maxNumberOfThreadsSuffix).println();
     }
 
     public ReporterClosable printAnalysis(AnalysisUniverse universe, Collection<String> libraries) {
@@ -534,7 +533,7 @@ public class ProgressReporter {
         if (numRuntimeCompiledMethods >= 0) {
             recordJsonMetric(ImageDetailKey.RUNTIME_COMPILED_METHODS_COUNT, numRuntimeCompiledMethods);
             l().a("%,9d ", numRuntimeCompiledMethods).doclink("runtime compiled methods", "#glossary-runtime-methods")
-                            .a(" (%.1f%% of all reachable methods)", Utils.toPercentage(numRuntimeCompiledMethods, reachableMethods), reachableMethods).println();
+                            .a(" (%.1f%% of all reachable methods)", ProgressReporterUtils.toPercentage(numRuntimeCompiledMethods, reachableMethods), reachableMethods).println();
         }
     }
 
@@ -592,7 +591,7 @@ public class ProgressReporter {
         stagePrinter.end(imageTimer.getTotalTime() + writeTimer.getTotalTime() + archiveTimer.getTotalTime());
         creationStageEndCompleted = true;
         String format = "%9s (%5.2f%%) for ";
-        l().a(format, ByteFormattingUtil.bytesToHuman(codeAreaSize), Utils.toPercentage(codeAreaSize, imageFileSize))
+        l().a(format, ByteFormattingUtil.bytesToHuman(codeAreaSize), ProgressReporterUtils.toPercentage(codeAreaSize, imageFileSize))
                         .doclink("code area", "#glossary-code-area").a(":%,10d compilation units", numCompilations).println();
         int numResources = 0;
         for (ConditionalRuntimeValue<ResourceStorageEntryBase> entry : Resources.currentLayer().resources().getValues()) {
@@ -601,16 +600,16 @@ public class ProgressReporter {
             }
         }
         recordJsonMetric(ImageDetailKey.IMAGE_HEAP_RESOURCE_COUNT, numResources);
-        l().a(format, ByteFormattingUtil.bytesToHuman(imageHeapSize), Utils.toPercentage(imageHeapSize, imageFileSize))
+        l().a(format, ByteFormattingUtil.bytesToHuman(imageHeapSize), ProgressReporterUtils.toPercentage(imageHeapSize, imageFileSize))
                         .doclink("image heap", "#glossary-image-heap").a(":%,9d objects and %,d resource%s", heapObjectCount, numResources, numResources == 1 ? "" : "s").println();
         long otherBytes = imageFileSize - codeAreaSize - imageHeapSize;
         if (debugInfoSize > 0) {
             recordJsonMetric(ImageDetailKey.DEBUG_INFO_SIZE, debugInfoSize); // Optional metric
-            DirectPrinter l = l().a(format, ByteFormattingUtil.bytesToHuman(debugInfoSize), Utils.toPercentage(debugInfoSize, imageFileSize))
+            DirectPrinter l = l().a(format, ByteFormattingUtil.bytesToHuman(debugInfoSize), ProgressReporterUtils.toPercentage(debugInfoSize, imageFileSize))
 
                             .doclink("debug info", "#glossary-debug-info");
             if (debugInfoTimer != null) {
-                l.a(" generated in %.1fs", Utils.millisToSeconds(debugInfoTimer.getTotalTime()));
+                l.a(" generated in %.1fs", ProgressReporterUtils.millisToSeconds(debugInfoTimer.getTotalTime()));
             }
             l.println();
             if (!(ImageSingletons.contains(NativeImageDebugInfoStripFeature.class) && ImageSingletons.lookup(NativeImageDebugInfoStripFeature.class).hasStrippedSuccessfully())) {
@@ -623,7 +622,7 @@ public class ProgressReporter {
         recordJsonMetric(ImageDetailKey.TOTAL_SIZE, imageFileSize);
         recordJsonMetric(ImageDetailKey.CODE_AREA_SIZE, codeAreaSize);
         recordJsonMetric(ImageDetailKey.NUM_COMP_UNITS, numCompilations);
-        l().a(format, ByteFormattingUtil.bytesToHuman(otherBytes), Utils.toPercentage(otherBytes, imageFileSize))
+        l().a(format, ByteFormattingUtil.bytesToHuman(otherBytes), ProgressReporterUtils.toPercentage(otherBytes, imageFileSize))
                         .doclink("other data", "#glossary-other-data").println();
         l().a("%9s in total image size", ByteFormattingUtil.bytesToHuman(imageFileSize));
         if (imageDiskFileSize >= 0) {
@@ -641,58 +640,13 @@ public class ProgressReporter {
         }
     }
 
-    private record BreakDownClassifier(Package javaPackage, Module javaModule, String location) {
-        static BreakDownClassifier of(Class<?> clazz) {
-            return new BreakDownClassifier(clazz.getPackage(), clazz.getModule(), sourcePath(clazz));
-        }
-
-        private static String sourcePath(Class<?> clazz) {
-            CodeSource codeSource = clazz.getProtectionDomain().getCodeSource();
-            if (codeSource != null && codeSource.getLocation() != null) {
-                String path = codeSource.getLocation().getPath();
-                // Use String API to determine basename of path to handle both / and \.
-                return path.substring(Math.max(path.lastIndexOf('/') + 1, path.lastIndexOf('\\') + 1));
-            }
-            return null;
-        }
-
-        public String renderToString(int maxLength) {
-            String locationSuffix = location == null ? "" : " (" + Utils.truncateName(location, 0.10) + ")";
-            String packageName = javaPackage == null ? "null" : javaPackage.getName();
-            String moduleNamePrefix = moduleNamePrefix(javaModule);
-            // Give remainder of space to package-part
-            int maxLengthPackage = maxLength - moduleNamePrefix.length() - locationSuffix.length();
-            return moduleNamePrefix + Utils.truncateFQN(packageName, maxLengthPackage) + locationSuffix;
-        }
-
-        public String[] elements() {
-            String moduleName = javaModule.isNamed() ? javaModule.getName() : "";
-            String packageName = javaPackage == null ? "" : javaPackage.getName();
-            String locationName = location == null ? "" : location;
-            return new String[]{mapToNativeImageRuntime(moduleName), packageName, locationName};
-        }
-    }
-
-    static String moduleNamePrefix(Module javaModule) {
-        String moduleName = javaModule.isNamed() ? javaModule.getName() : "";
-        return Utils.truncateFQN(mapToNativeImageRuntime(moduleName), 0.10) + "/";
-    }
-
-    private static String mapToNativeImageRuntime(String moduleName) {
-        String modulePrefix = "org.graalvm.nativeimage.";
-        if (moduleName.equals(modulePrefix + "builder")) {
-            return modulePrefix + "runtime";
-        }
-        return moduleName;
-    }
-
     private void printBreakdowns() {
-        Map<BreakDownClassifier, Long> codeBreakdown = CodeBreakdownProvider.getAndClear().entrySet().stream()
+        Map<ProgressReporterUtils.BreakDownClassifier, Long> codeBreakdown = CodeBreakdownProvider.getAndClear().entrySet().stream()
                         .collect(Collectors.groupingBy(
-                                        entry -> BreakDownClassifier.of(entry.getKey()),
-                                        Collectors.summingLong(entry -> entry.getValue())));
+                                        entry -> ProgressReporterUtils.BreakDownClassifier.of(entry.getKey()),
+                                        Collectors.summingLong(Entry::getValue)));
 
-        List<Entry<BreakDownClassifier, Long>> sortedBreakdownData = codeBreakdown.entrySet().stream()
+        List<Entry<ProgressReporterUtils.BreakDownClassifier, Long>> sortedBreakdownData = codeBreakdown.entrySet().stream()
                         .sorted(Entry.comparingByValue(Comparator.reverseOrder())).toList();
 
         if (SubstrateOptions.BuildOutputCodeBreakdownFile.getValue()) {
@@ -725,14 +679,14 @@ public class ProgressReporter {
         long printedHeapBytes = 0;
         long printedCodeItems = 0;
         long printedHeapItems = 0;
-        Iterator<Entry<BreakDownClassifier, Long>> packagesBySize = sortedBreakdownData.iterator();
+        Iterator<Entry<ProgressReporterUtils.BreakDownClassifier, Long>> packagesBySize = sortedBreakdownData.iterator();
         HeapBreakdownProvider heapBreakdown = HeapBreakdownProvider.singleton();
         Iterator<HeapBreakdownProvider.HeapBreakdownEntry> typesBySizeInHeap = heapBreakdown.getSortedBreakdownEntries().iterator();
         for (int i = 0; i < MAX_NUM_BREAKDOWN; i++) {
             String codeSizePart = "";
             /* <- 10% for module name -><- remainder for class FQN -><- 10% for location -> */
             if (packagesBySize.hasNext()) {
-                Entry<BreakDownClassifier, Long> entry = packagesBySize.next();
+                Entry<ProgressReporterUtils.BreakDownClassifier, Long> entry = packagesBySize.next();
                 String sizeStr = String.format("%9s ", ByteFormattingUtil.bytesToHuman(entry.getValue()));
                 String entryStr = entry.getKey().renderToString(p.middle() - sizeStr.length());
                 codeSizePart = sizeStr + entryStr;
@@ -778,7 +732,7 @@ public class ProgressReporter {
         l().printLineSeparator();
         l().yellowBold().a("Recommendations:").reset().println();
         for (UserRecommendation r : topApplicableRecommendations) {
-            String alignment = Utils.stringFilledWith(Math.max(1, 5 - r.id().length()), " ");
+            String alignment = ProgressReporterUtils.stringFilledWith(Math.max(1, 5 - r.id().length()), " ");
             l().a(" ").doclink(r.id(), "#recommendation-" + r.id().toLowerCase(Locale.ROOT)).a(":").a(alignment).a(r.description()).println();
         }
     }
@@ -811,7 +765,7 @@ public class ProgressReporter {
         l().printLineSeparator();
         printResourceStatistics();
 
-        double totalSeconds = Utils.millisToSeconds(getTimer(TimerCollection.Registry.TOTAL).getTotalTime());
+        double totalSeconds = ProgressReporterUtils.millisToSeconds(getTimer(TimerCollection.Registry.TOTAL).getTotalTime());
         recordJsonMetric(ResourceUsageKey.TOTAL_SECS, totalSeconds);
 
         createAdditionalArtifacts(imageName, generator, buildOutcome.successful(), parsedHostedOptions);
@@ -914,9 +868,9 @@ public class ProgressReporter {
     }
 
     private void printResourceStatistics() {
-        double totalProcessTimeSeconds = Utils.millisToSeconds(ManagementFactory.getRuntimeMXBean().getUptime());
+        double totalProcessTimeSeconds = ProgressReporterUtils.millisToSeconds(ManagementFactory.getRuntimeMXBean().getUptime());
         GCStats gcStats = GCStats.getCurrent();
-        double gcSeconds = Utils.millisToSeconds(gcStats.totalTimeMillis);
+        double gcSeconds = ProgressReporterUtils.millisToSeconds(gcStats.totalTimeMillis);
         recordJsonMetric(ResourceUsageKey.GC_COUNT, gcStats.totalCount);
         recordJsonMetric(ResourceUsageKey.GC_SECS, gcSeconds);
         CenteredTextPrinter p = centered();
@@ -930,7 +884,7 @@ public class ProgressReporter {
         long processCPUTime = getOperatingSystemMXBean().getProcessCpuTime();
         double cpuLoad = UNAVAILABLE_METRIC;
         if (processCPUTime > 0) {
-            cpuLoad = Utils.nanosToSeconds(processCPUTime) / totalProcessTimeSeconds;
+            cpuLoad = ProgressReporterUtils.nanosToSeconds(processCPUTime) / totalProcessTimeSeconds;
             p.a(" | ").doclink("CPU load", "#glossary-cpu-load").a(": ").a("%.2f", cpuLoad);
         }
         recordJsonMetric(ResourceUsageKey.CPU_LOAD, cpuLoad);
@@ -947,7 +901,7 @@ public class ProgressReporter {
         if (gcTimeDeltaMillis > EXCESSIVE_GC_MIN_THRESHOLD_MILLIS && ratio > EXCESSIVE_GC_RATIO) {
             l().redBold().a("GC warning").reset()
                             .a(": %.1fs spent in %d GCs during the last stage, taking up %.2f%% of the time.",
-                                            Utils.millisToSeconds(gcTimeDeltaMillis), currentGCStats.totalCount - lastGCStats.totalCount, ratio * 100)
+                                            ProgressReporterUtils.millisToSeconds(gcTimeDeltaMillis), currentGCStats.totalCount - lastGCStats.totalCount, ratio * 100)
                             .println();
             l().a("            Please ensure more than %s of memory is available for Native Image", ByteFormattingUtil.bytesToHuman(ProgressReporterCHelper.getPeakRSS())).println();
             l().a("            to reduce GC overhead and improve image build time.").println();
@@ -985,81 +939,6 @@ public class ProgressReporter {
 
     private static OperatingSystemMXBean getOperatingSystemMXBean() {
         return (OperatingSystemMXBean) ManagementFactory.getOperatingSystemMXBean();
-    }
-
-    static final class Utils {
-        private static final double MILLIS_TO_SECONDS = 1000d;
-        private static final double NANOS_TO_SECONDS = 1000d * 1000d * 1000d;
-        public static final String TRUNCATION_PLACEHOLDER = "~";
-
-        private static double millisToSeconds(double millis) {
-            return millis / MILLIS_TO_SECONDS;
-        }
-
-        private static double nanosToSeconds(double nanos) {
-            return nanos / NANOS_TO_SECONDS;
-        }
-
-        private static String getUsedMemory() {
-            return ByteFormattingUtil.bytesToHumanGB(Runtime.getRuntime().totalMemory() - Runtime.getRuntime().freeMemory());
-        }
-
-        private static String stringFilledWith(int size, String fill) {
-            return fill.repeat(size);
-        }
-
-        private static double toPercentage(long part, long total) {
-            return part / (double) total * 100;
-        }
-
-        private static String truncateName(String name, double maxLineRatio) {
-            int length = name.length();
-            int maxLength = maxLength(maxLineRatio);
-            if (length <= maxLength) {
-                return name;
-            }
-            return TRUNCATION_PLACEHOLDER + name.substring(length - maxLength + TRUNCATION_PLACEHOLDER.length(), length);
-        }
-
-        static String truncateFQN(String fqn, double maxLineRatio) {
-            return truncateFQN(fqn, maxLength(maxLineRatio));
-        }
-
-        static int maxLength(double maxLineRatio) {
-            return (int) Math.floor(CHARACTERS_PER_LINE * maxLineRatio);
-        }
-
-        private static String truncateFQN(String fqn, int maxLength) {
-            int classNameLength = fqn.length();
-            if (classNameLength <= maxLength) {
-                return fqn;
-            }
-            StringBuilder sb = new StringBuilder();
-            int currentDot = -1;
-            while (true) {
-                int nextDot = fqn.indexOf('.', currentDot + 1);
-                if (nextDot < 0) { // Not more dots, handle the rest and return.
-                    String rest = fqn.substring(currentDot + 1);
-                    int sbLength = sb.length();
-                    int restLength = rest.length();
-                    if (sbLength + restLength <= maxLength) {
-                        sb.append(rest);
-                    } else {
-                        int remainingSpaceDivBy2 = (maxLength - sbLength) / 2;
-                        sb.append(rest, 0, remainingSpaceDivBy2 - 1).append(TRUNCATION_PLACEHOLDER).append(rest, restLength - remainingSpaceDivBy2, restLength);
-                    }
-                    break;
-                }
-                sb.append(fqn.charAt(currentDot + 1)).append('.');
-                if (sb.length() + (classNameLength - nextDot) <= maxLength) {
-                    // Rest fits maxLength, append and return.
-                    sb.append(fqn.substring(nextDot + 1));
-                    break;
-                }
-                currentDot = nextDot;
-            }
-            return sb.toString();
-        }
     }
 
     private record GCStats(long totalCount, long totalTimeMillis) {
@@ -1318,7 +1197,7 @@ public class ProgressReporter {
         }
 
         final String progressBarStartPadding() {
-            return Utils.stringFilledWith(PROGRESS_BAR_START - getCurrentTextLength(), " ");
+            return ProgressReporterUtils.stringFilledWith(PROGRESS_BAR_START - getCurrentTextLength(), " ");
         }
 
         void reportProgress() {
@@ -1338,11 +1217,11 @@ public class ProgressReporter {
                 a("]").reset();
             }
 
-            String suffix = String.format("(%.1fs @ %s)", Utils.millisToSeconds(totalTime), Utils.getUsedMemory());
+            String suffix = String.format("(%.1fs @ %s)", ProgressReporterUtils.millisToSeconds(totalTime), ProgressReporterUtils.getUsedMemory());
             int textLength = getCurrentTextLength();
             // TODO: `assert textLength > 0;` should be used here but tests do not start stages
             // properly (GR-35721)
-            String padding = Utils.stringFilledWith(Math.max(0, CHARACTERS_PER_LINE - textLength - suffix.length()), " ");
+            String padding = ProgressReporterUtils.stringFilledWith(Math.max(0, CHARACTERS_PER_LINE - textLength - suffix.length()), " ");
             a(padding).dim().a(suffix).reset().flushln();
 
             activeBuildStage = null;
@@ -1461,7 +1340,7 @@ public class ProgressReporter {
         TwoColumnPrinter jumpToMiddle() {
             int remaining = middle() - getCurrentTextLength();
             assert remaining >= 0 : "Column text too wide";
-            a(Utils.stringFilledWith(remaining, " "));
+            a(ProgressReporterUtils.stringFilledWith(remaining, " "));
             assert getCurrentTextLength() == middle();
             return this;
         }
@@ -1479,7 +1358,7 @@ public class ProgressReporter {
 
         @Override
         public void flushln() {
-            String padding = Utils.stringFilledWith((Math.max(0, CHARACTERS_PER_LINE - getCurrentTextLength())) / 2, " ");
+            String padding = ProgressReporterUtils.stringFilledWith((Math.max(0, CHARACTERS_PER_LINE - getCurrentTextLength())) / 2, " ");
             print(padding);
             super.flushln();
         }
